Improve Bedrock model ordering and labels
